refactor: `UIAlertController` callings from Extensions, add array of summaries
